+++ src/types.ts
@@ -1,4 +1,13 @@
-import { Store } from 'effector';
+import { Event, Store } from 'effector';
+
+export interface WithPersistentFlushEvent {
+  /** Identifier for current flush flow */
+  readonly id: symbol;
+}
+export interface WithPersistentFlushFailEvent extends WithPersistentFlushEvent {
+  /** Reason of the failure */
+  readonly error: unknown;
+}
 
 export interface WithPersistentOptions<
   State = any,
@@ -12,7 +21,28 @@ export interface WithPersistentOptions<
    */
   flushDelay?: number;
   /**
-   * Optional flag store to disable writes to storage.
+   * An Event to trigger before flushing to driver. An `id` in payload can be
+   * used in all the rest "flush" events to identify the flow when it's needed.
+   */
+  onFlushStart?: Event<WithPersistentFlushEvent>;
+  /**
+   * An Event to trigger when flush succeeds. An `id` in payload refers to `id`
+   * from appropriate `onFlushStart` payload.
+   */
+  onFlushDone?: Event<WithPersistentFlushEvent>;
+  /**
+   * An Event to trigger when flush fails. An `id` in payload refers to `id`
+   * from appropriate `onFlushStart` payload.
+   */
+  onFlushFail?: Event<WithPersistentFlushFailEvent>;
+  /**
+   * An Event to trigger before flushing to driver. This al always triggering
+   * after either `onFlushDone` or `onFlushFail`. An `id` in payload refers to
+   * `id` from appropriate `onFlushStart` payload.
+   */
+  onFlushFinally?: Event<WithPersistentFlushEvent>;
+  /**
+   * A `filter` Store to disable writes to Driver.
    */
   readOnly?: Store<boolean>;
   /**